+++ Doolda/Doolda/Domain/UseCases/EditPageUseCase.swift
@@ -21,28 +21,7 @@ enum EditPageUseCaseError: LocalizedError {
     }
 }
 
-protocol EditPageUseCaseProtocol {
-    var selectedComponentPublisher: Published<ComponentEntity?>.Publisher { get }
-    var rawPagePublisher: Published<RawPageEntity?>.Publisher { get }
-    var errorPublisher: Published<Error?>.Publisher { get }
-    var resultPublisher: Published<Bool?>.Publisher { get }
-    
-    func selectComponent(at point: CGPoint)
-    func moveComponent(to point: CGPoint)
-    func rotateComponent(by angle: CGFloat)
-    func scaleComponent(by scale: CGFloat)
-    func bringComponentFront()
-    func sendComponentBack()
-    func removeComponent()
-    func addComponent(_ component: ComponentEntity)
-    
-    func changeTextComponent(into content: TextComponentEntity)
-    
-    func changeBackgroundType(_ backgroundType: BackgroundType)
-    func savePage(author: User, metaData: PageEntity?)
-}
-
-class EditPageUseCase: EditPageUseCaseProtocol {
+final class EditPageUseCase: EditPageUseCaseProtocol {
     var selectedComponentPublisher: Published<ComponentEntity?>.Publisher { self.$selectedComponent }
     var rawPagePublisher: Published<RawPageEntity?>.Publisher { self.$rawPage }
     var errorPublisher: Published<Error?>.Publisher { self.$error }
